Moved CSS files to CSS folder
