+++ Frontend/src/components/ForgotPassword.js
@@ -11,7 +11,7 @@ import {
 import { useForm } from "react-hook-form";
 import axios from "axios";
 import KeyStore from "../KeyStore";
-import "./Login.css";
+import "../CSS/Login.css";
 import { useNavigate } from "react-router-dom";
 
 const ForgotPassword = () => {
