+++ Frontend/src/components/UserCard.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./Post.css";
+import "../CSS/Post.css";
 import Avatar from "../Utility/BackgroundLetterAvatars";
 import { Button } from "@mui/material";
 import { Link } from "react-router-dom";
