+++ Frontend/src/components/UserTweets.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { Box, IconButton } from "@mui/material";
 import Post from "./Post";
-import "./Feed.css";
+import "../CSS/Feed.css";
 import KeyStore from "../KeyStore";
 import axios from "axios";
 import FlipMove from "react-flip-move";
