+++ Frontend/src/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import "./index.css";
+import "./CSS/index.css";
 import { BrowserRouter } from "react-router-dom";
 import App from "./App";
